# api/openapi/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.BatchFavoriteStatusRequest:
    properties:
      video_ids:
        items:
          type: integer
        maxItems: 100
        minItems: 1
        type: array
    required:
    - video_ids
    type: object
  dto.BatchFollowStatusRequest:
    properties:
      user_ids:
        items:
          type: integer
        maxItems: 100
        minItems: 1
        type: array
    required:
    - user_ids
    type: object
...
  response.ErrorInfo:
    properties:
      code:
        type: integer
      message:
        type: string
      type:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorInfo'
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: 127.0.0.1:8000
info:
  contact:
    email: support@vida.com
    name: API Support
  description: 视频分享平台 API 服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Vida-Go API
  version: "1.0"
paths:
...
  /favorites/{video_id}:
    delete:
      description: 取消对指定视频的点赞
      parameters:
      - description: 视频ID
        in: path
        name: video_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 取消点赞成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 未点赞
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 取消点赞
      tags:
      - 点赞
    post:
      description: 对指定视频点赞
      parameters:
      - description: 视频ID
        in: path
        name: video_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 点赞成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 已点赞
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 视频不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 点赞视频
      tags:
      - 点赞
  /favorites/{video_id}/status:
    get:
      description: 查询是否点赞了指定视频
      parameters:
      - description: 视频ID
        in: path
        name: video_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取点赞状态
      tags:
      - 点赞
  /favorites/batch/status:
    post:
      consumes:
      - application/json
      description: 批量查询对多个视频的点赞状态
      parameters:
      - description: 视频ID列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BatchFavoriteStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 批量查询点赞状态
      tags:
      - 点赞
  /favorites/my/list:
    get:
      description: 获取当前用户的点赞记录列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取我的点赞列表
      tags:
      - 点赞
  /favorites/my/videos:
    get:
      description: 获取当前用户点赞过的视频ID列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取我点赞的视频ID列表
      tags:
      - 点赞
  /favorites/video/{video_id}/list:
    get:
      description: 获取指定视频的点赞用户列表
      parameters:
      - description: 视频ID
        in: path
        name: video_id
        required: true
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取视频点赞列表
      tags:
      - 点赞
  /relations/batch/status:
    post:
      consumes:
      - application/json
      description: 批量查询对多个用户的关注状态
      parameters:
      - description: 用户ID列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BatchFollowStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 批量查询关注状态
      tags:
      - 关注
  /relations/follow/{id}:
    post:
      description: 关注指定用户
      parameters:
      - description: 被关注用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 关注成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 不能关注自己/已关注
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 关注用户
      tags:
      - 关注
  /relations/followers/{id}:
    get:
      description: 获取指定用户的粉丝列表
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户粉丝列表
      tags:
      - 关注
  /relations/followers/my/list:
    get:
      description: 获取当前用户的粉丝列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取我的粉丝列表
      tags:
      - 关注
  /relations/following/{id}:
    get:
      description: 获取指定用户的关注列表
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户关注列表
      tags:
      - 关注
  /relations/following/{id}/status:
    get:
      description: 查询是否关注了指定用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取关注状态
      tags:
      - 关注
  /relations/following/my/list:
    get:
      description: 获取当前用户的关注列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取我的关注列表
      tags:
      - 关注
  /relations/mutual:
    get:
      description: 获取与当前用户互相关注的用户列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取互相关注列表
      tags:
      - 关注
  /relations/unfollow/{id}:
    post:
      description: 取消关注指定用户
      parameters:
      - description: 被取消关注用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 取消关注成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 未关注该用户
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 取消关注
      tags:
      - 关注
...
securityDefinitions:
  BearerAuth:
    description: '输入格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"